requirements.txt
bcrypt                 4.3.0
blinker                1.9.0
click                  8.1.8
colorama               0.4.6
dnspython              2.7.0
email_validator        2.2.0
Flask                  3.1.0
Flask-Bcrypt           1.0.1
Flask-Login            0.6.3
Flask-SQLAlchemy       3.1.1
Flask-WTF              1.2.2
greenlet               3.2.1
idna                   3.10
itsdangerous           2.2.0
Jinja2                 3.1.6
MarkupSafe             3.0.2
mysql-connector-python 9.3.0
pip                    24.2
PyMySQL                1.1.1
python-dotenv          1.1.0
SQLAlchemy             2.0.40
typing_extensions      4.13.2
Werkzeug               3.1.3
WTForms                3.2.1
